requirements.txt
flask==2.3.3
openai==1.3.0
playwright==1.38.0
python-dotenv==1.0.0
python-docx==0.8.11
pdfminer.six==20221105
pandas==2.0.0
beautifulsoup4==4.11.1
lxml==4.9.2
gunicorn==20.1.0
requests==2.31.0
